revert: restore the single GET /info bootstrap payload (owner decision)
Undo the B6 per-resource split and the subsequent info-package dissolution:
the user-facing API is read-only, so one GET-only bootstrap payload grouped
under info is the right shape after all. The /user/me, /group, /service,
/connection, /notice, /ticket, /request GETs are gone; GET /info again returns
{user, group, user_list, service, connection, notice, ticket, request, info},
byte-identical to the pre-refactor wire (verified against the preserved
pre-B6 golden blob — the interim wire edits to user/stripe_customer_id and
JPNIC handles are rolled back too).

Kept from the interim work: middleware.UserAuth/CurrentUser, the
CanViewGroup/IsPaidMemberType helpers, core.ServiceCode/ConnectionCode, and
the deletion of the dead legacy wire types. The info package now holds the
DTOs (interface.go) and the single blob handler (v0/info.go); the per-domain
user-view types and handlers are removed.

Golden: the seven per-resource cases/files are gone; user_info is the full
blob again and the suite is green.
