+++ pkg/api/core/group/service/interface.go
@@ -28,64 +28,6 @@ type Input struct {
 	BGPComment     string           `json:"bgp_comment"`
 }
 
-// JPNIC is a JPNIC contact as shown to users. The admin contact omits
-// its address fields on the wire.
-type JPNIC struct {
-	ID          uint   `json:"id"`
-	JPNICHandle string `json:"jpnic_handle"`
-	Name        string `json:"name"`
-	NameEn      string `json:"name_en"`
-	Mail        string `json:"mail"`
-	Org         string `json:"org"`
-	OrgEn       string `json:"org_en"`
-	PostCode    string `json:"postcode"`
-	Address     string `json:"address"`
-	AddressEn   string `json:"address_en"`
-	Dept        string `json:"dept"`
-	DeptEn      string `json:"dept_en"`
-	Tel         string `json:"tel"`
-	Fax         string `json:"fax"`
-	Country     string `json:"country"`
-}
-
-// Service is the user-facing wire shape of an enabled service
-// (GET /service), including the capability flags from the type registry.
-type Service struct {
-	ID             uint    `json:"id"`
-	ServiceID      string  `json:"service_id"`
-	ServiceType    string  `json:"service_type"`
-	NeedRoute      bool    `json:"need_route"`
-	NeedBGP        bool    `json:"need_bgp"`
-	NeedJPNIC      bool    `json:"need_jpnic"`
-	AddAllow       bool    `json:"add_allow"`
-	Pass           bool    `json:"pass"`
-	Org            string  `json:"org"`
-	OrgEn          string  `json:"org_en"`
-	PostCode       string  `json:"postcode"`
-	Address        string  `json:"address"`
-	AddressEn      string  `json:"address_en"`
-	ASN            *uint   `json:"asn"`
-	AveUpstream    uint    `json:"avg_upstream"`
-	MaxUpstream    uint    `json:"max_upstream"`
-	AveDownstream  uint    `json:"avg_downstream"`
-	MaxDownstream  uint    `json:"max_downstream"`
-	MaxBandWidthAS string  `json:"max_bandwidth_as"`
-	JPNICAdmin     JPNIC   `json:"jpnic_admin"`
-	JPNICTech      []JPNIC `json:"jpnic_tech"`
-	IP             []IP    `json:"ip"`
-}
-
-// IP is an opened IP assignment as shown to users.
-type IP struct {
-	ID        uint   `json:"id"`
-	Version   uint   `json:"version"`
-	Name      string `json:"name"`
-	IP        string `json:"ip"`
-	Plan      []Plan `json:"plan" `
-	PlanJPNIC string `json:"" gorm:"size:65535"`
-	UseCase   string `json:"use_case"`
-}
-
 type IPInput struct {
 	Version   uint         `json:"version"`
 	Name      string       `json:"name"`
@@ -108,13 +50,4 @@ type ResultOne struct {
 type ResultDatabase struct {
 	Err     error
 	Service []core.Service
-}
-
-type Plan struct {
-	ID       uint   `json:"id"`
-	IPID     uint   `json:"ip_id"`
-	Name     string `json:"name"`
-	After    uint   `json:"after"`
-	HalfYear uint   `json:"half_year"`
-	OneYear  uint   `json:"one_year"`
 }
