+++ pkg/api/core/user/interface.go
@@ -1,8 +1,6 @@
 package user
 
 import (
-	"time"
-
 	"github.com/homenoc/dsbd-backend/pkg/api/core"
 )
 
@@ -14,42 +12,8 @@ type Input struct {
 	Level  uint   `json:"level"`
 }
 
-// User is the user-facing wire shape of an account (GET /user/me and the
-// group member list in GET /group).
-type User struct {
-	ID                uint       `json:"id"`
-	GroupID           uint       `json:"group_id"`
-	Name              string     `json:"name"`
-	NameEn            string     `json:"name_en"`
-	Email             string     `json:"email"`
-	Status            uint       `json:"status"`
-	Level             uint       `json:"level"`
-	MailVerify        *bool      `json:"mail_verify"`
-	AntisocialCheck   *bool      `json:"antisocial_check"`
-	AntisocialCheckAt *time.Time `json:"antisocial_check_at"`
-}
-
-// NewUser projects a core.User onto its user-facing wire shape.
-func NewUser(u core.User) User {
-	var groupID uint
-	if u.GroupID != nil {
-		groupID = *u.GroupID
-	}
-	return User{
-		ID:                u.ID,
-		GroupID:           groupID,
-		Name:              u.Name,
-		NameEn:            u.NameEn,
-		Email:             u.Email,
-		Level:             u.Level,
-		MailVerify:        u.MailVerify,
-		AntisocialCheck:   u.AntisocialCheck,
-		AntisocialCheckAt: u.AntisocialCheckAt,
-	}
-}
-
 type Result struct {
-	User []User `json:"user"`
+	User []core.User `json:"user"`
 }
 
 type ResultAdmin struct {
