--- pkg/api/testdata/golden/user_connection.json
@@ -1,12 +1,0 @@
-{
-  "status": 200,
-  "body": {
-    "connection": [
-      {
-        "id": 1,
-        "open": false,
-        "service_id": "1-3B00001-EIP001"
-      }
-    ]
-  }
-}
