+++ pkg/api/testdata/golden/user_info.json
@@ -1,6 +1,120 @@
 {
   "status": 200,
   "body": {
-    "info": null
+    "connection": [
+      {
+        "id": 1,
+        "open": false,
+        "service_id": "1-3B00001-EIP001"
+      }
+    ],
+    "group": {
+      "add_allow": true,
+      "address": "東京都千代田区",
+      "address_en": "Chiyoda-ku, Tokyo",
+      "agree": true,
+      "contract": "",
+      "country": "Japan",
+      "coupon_id": "",
+      "expired_status": 0,
+      "id": 1,
+      "is_expired": true,
+      "is_stripe_id": false,
+      "member_expired": null,
+      "member_type": "一般会員",
+      "member_type_id": 1,
+      "org": "テスト組織",
+      "org_en": "Test Organization",
+      "pass": true,
+      "postcode": "100-0001",
+      "question": "開発用テストグループです",
+      "tel": "03-1234-5678"
+    },
+    "info": null,
+    "notice": null,
+    "request": null,
+    "service": [
+      {
+        "add_allow": true,
+        "address": "東京都千代田区",
+        "address_en": "Chiyoda-ku, Tokyo",
+        "asn": null,
+        "avg_downstream": 100,
+        "avg_upstream": 100,
+        "id": 1,
+        "ip": null,
+        "jpnic_admin": {
+          "address": "",
+          "address_en": "",
+          "country": "",
+          "dept": "",
+          "dept_en": "",
+          "fax": "",
+          "id": 0,
+          "mail": "",
+          "name": "",
+          "name_en": "",
+          "org": "",
+          "org_en": "",
+          "postcode": "",
+          "tel": ""
+        },
+        "jpnic_tech": null,
+        "max_bandwidth_as": "",
+        "max_downstream": 1000,
+        "max_upstream": 1000,
+        "need_bgp": true,
+        "need_jpnic": true,
+        "need_route": true,
+        "org": "テスト組織",
+        "org_en": "Test Organization",
+        "pass": true,
+        "postcode": "100-0001",
+        "service_id": "1-3B00001",
+        "service_type": "3B00"
+      }
+    ],
+    "ticket": null,
+    "user": {
+      "antisocial_check": false,
+      "antisocial_check_at": null,
+      "email": "master@example.com",
+      "group_id": 1,
+      "id": 1,
+      "level": 1,
+      "mail_verify": true,
+      "name": "マスターユーザー",
+      "name_en": "Master User",
+      "status": 0,
+      "stripe_customer_id": ""
+    },
+    "user_list": [
+      {
+        "antisocial_check": false,
+        "antisocial_check_at": null,
+        "email": "master@example.com",
+        "group_id": 1,
+        "id": 1,
+        "level": 1,
+        "mail_verify": true,
+        "name": "マスターユーザー",
+        "name_en": "Master User",
+        "status": 0,
+        "stripe_customer_id": ""
+      },
+      {
+        "antisocial_check": false,
+        "antisocial_check_at": null,
+        "email": "member@example.com",
+        "group_id": 1,
+        "id": 2,
+        "level": 2,
+        "mail_verify": true,
+        "name": "一般ユーザー",
+        "name_en": "Member User",
+        "status": 0,
+        "stripe_customer_id": ""
+      }
+    ]
   }
 }
