--- pkg/api/testdata/golden/user_notice.json
@@ -1,6 +1,0 @@
-{
-  "status": 200,
-  "body": {
-    "notice": null
-  }
-}
