--- pkg/api/testdata/golden/user_request.json
@@ -1,6 +1,0 @@
-{
-  "status": 200,
-  "body": {
-    "request": null
-  }
-}
