--- pkg/api/testdata/golden/user_service.json
@@ -1,46 +1,0 @@
-{
-  "status": 200,
-  "body": {
-    "service": [
-      {
-        "add_allow": true,
-        "address": "東京都千代田区",
-        "address_en": "Chiyoda-ku, Tokyo",
-        "asn": null,
-        "avg_downstream": 100,
-        "avg_upstream": 100,
-        "id": 1,
-        "ip": null,
-        "jpnic_admin": {
-          "address": "",
-          "address_en": "",
-          "country": "",
-          "dept": "",
-          "dept_en": "",
-          "fax": "",
-          "id": 0,
-          "mail": "",
-          "name": "",
-          "name_en": "",
-          "org": "",
-          "org_en": "",
-          "postcode": "",
-          "tel": ""
-        },
-        "jpnic_tech": null,
-        "max_bandwidth_as": "",
-        "max_downstream": 1000,
-        "max_upstream": 1000,
-        "need_bgp": true,
-        "need_jpnic": true,
-        "need_route": true,
-        "org": "テスト組織",
-        "org_en": "Test Organization",
-        "pass": true,
-        "postcode": "100-0001",
-        "service_id": "1-3B00001",
-        "service_type": "3B00"
-      }
-    ]
-  }
-}
