--- pkg/api/testdata/golden/user_ticket.json
@@ -1,6 +1,0 @@
-{
-  "status": 200,
-  "body": {
-    "ticket": null
-  }
-}
